fix: refactor routes to dialogs
