+++ frontend/src/components/copy-with-tooltip/copy-with-tooltip.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import CopyToClipboard from 'react-copy-to-clipboard'
+import { Copy } from '../icons/copy'
 
 import { Tooltip } from '../tooltip'
 
 interface CopyWithtooltipProps {
-  children: React.ReactElement
+  children?: React.ReactElement
   text: string
 }
 
@@ -29,7 +30,16 @@ export function CopyWithTooltip({ children, text }: CopyWithtooltipProps) {
     <CopyToClipboard text={text} onCopy={() => setCopied(true)}>
       {/* Needs this wrapping div as tooltip component interfers with element used to capture click for copy */}
       <span>
-        <Tooltip trigger={children} content='Copied' isOpen={copied} />
+        <Tooltip
+          trigger={(
+            <span style={{ cursor: 'pointer' }}>
+              {children}
+              <Copy style={{ width: 13, marginLeft: 6 }} />
+            </span>
+          )}
+          content='Copied'
+          isOpen={copied}
+        />
       </span>
     </CopyToClipboard>
   )
