+++ frontend/src/components/delete-wallet-dialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './delete-wallet-dialog'
