+++ frontend/src/components/sign-message-dialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './sign-message-dialog'
