+++ frontend/src/components/taint-key-dialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './taint-key-dialog'
