+++ frontend/src/components/update-keypair-dialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './update-keypair-dialog'
