+++ frontend/src/wailsjs/go/models.ts
@@ -1,132 +1,31 @@
-export namespace jsonrpc {
-	
-	export class Request {
-	    jsonrpc: string;
-	    method: string;
-	    params?: any;
-	    id?: string;
-	
-	    static createFrom(source: any = {}) {
-	        return new Request(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.jsonrpc = source["jsonrpc"];
-	        this.method = source["method"];
-	        this.params = source["params"];
-	        this.id = source["id"];
-	    }
-	}
-	export class ErrorDetails {
-	    code: number;
-	    message: string;
-	    data?: string;
-	
-	    static createFrom(source: any = {}) {
-	        return new ErrorDetails(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.code = source["code"];
-	        this.message = source["message"];
-	        this.data = source["data"];
-	    }
-	}
-	export class Response {
-	    jsonrpc: string;
-	    result?: any;
-	    error?: ErrorDetails;
-	    id?: string;
-	
-	    static createFrom(source: any = {}) {
-	        return new Response(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.jsonrpc = source["jsonrpc"];
-	        this.result = source["result"];
-	        this.error = this.convertValues(source["error"], ErrorDetails);
-	        this.id = source["id"];
-	    }
-	
-		convertValues(a: any, classs: any, asMap: boolean = false): any {
-		    if (!a) {
-		        return a;
-		    }
-		    if (a.slice) {
-		        return (a as any[]).map(elem => this.convertValues(elem, classs));
-		    } else if ("object" === typeof a) {
-		        if (asMap) {
-		            for (const key of Object.keys(a)) {
-		                a[key] = new classs(a[key]);
-		            }
-		            return a;
-		        }
-		        return new classs(a);
-		    }
-		    return a;
-		}
-	}
-
-}
-
 export namespace backend {
 	
-	export class GetVersionResponse {
+	export class CheckVersionResponse {
 	    version: string;
-	    gitHash: string;
+	    releaseUrl: string;
 	
 	    static createFrom(source: any = {}) {
-	        return new GetVersionResponse(source);
+	        return new CheckVersionResponse(source);
 	    }
 	
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
 	        this.version = source["version"];
-	        this.gitHash = source["gitHash"];
+	        this.releaseUrl = source["releaseUrl"];
 	    }
 	}
-	export class InitialiseAppRequest {
-	    vegaHome: string;
+	export class ConsentToTransactionRequest {
+	    txId: string;
+	    decision: boolean;
 	
 	    static createFrom(source: any = {}) {
-	        return new InitialiseAppRequest(source);
+	        return new ConsentToTransactionRequest(source);
 	    }
 	
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.vegaHome = source["vegaHome"];
-	    }
-	}
-	export class Interaction {
-	    traceId: string;
-	    type: string;
-	    content: any;
-	
-	    static createFrom(source: any = {}) {
-	        return new Interaction(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.traceId = source["traceId"];
-	        this.type = source["type"];
-	        this.content = source["content"];
-	    }
-	}
-	export class StartServiceRequest {
-	    network: string;
-	
-	    static createFrom(source: any = {}) {
-	        return new StartServiceRequest(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.network = source["network"];
+	        this.txId = source["txId"];
+	        this.decision = source["decision"];
 	    }
 	}
 	export class ConsentRequest {
@@ -164,6 +63,22 @@ export namespace backend {
 		    return a;
 		}
 	}
+	export class Interaction {
+	    traceId: string;
+	    type: string;
+	    content: any;
+	
+	    static createFrom(source: any = {}) {
+	        return new Interaction(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.traceId = source["traceId"];
+	        this.type = source["type"];
+	        this.content = source["content"];
+	    }
+	}
 	export class ClearSentTransactionRequest {
 	    txId: string;
 	
@@ -174,20 +89,6 @@ export namespace backend {
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
 	        this.txId = source["txId"];
-	    }
-	}
-	export class ConsentToTransactionRequest {
-	    txId: string;
-	    decision: boolean;
-	
-	    static createFrom(source: any = {}) {
-	        return new ConsentToTransactionRequest(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.txId = source["txId"];
-	        this.decision = source["decision"];
 	    }
 	}
 	export class GetConsentRequestRequest {
@@ -214,6 +115,32 @@ export namespace backend {
 	        if ('string' === typeof source) source = JSON.parse(source);
 	        this.url = source["url"];
 	        this.running = source["running"];
+	    }
+	}
+	export class GetVersionResponse {
+	    version: string;
+	    gitHash: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new GetVersionResponse(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.version = source["version"];
+	        this.gitHash = source["gitHash"];
+	    }
+	}
+	export class InitialiseAppRequest {
+	    vegaHome: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new InitialiseAppRequest(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.vegaHome = source["vegaHome"];
 	    }
 	}
 	export class ListConsentRequestsResponse {
@@ -329,18 +256,16 @@ export namespace backend {
 	        this.networks = source["networks"];
 	    }
 	}
-	export class CheckVersionResponse {
-	    version: string;
-	    releaseUrl: string;
+	export class StartServiceRequest {
+	    network: string;
 	
 	    static createFrom(source: any = {}) {
-	        return new CheckVersionResponse(source);
+	        return new StartServiceRequest(source);
 	    }
 	
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.version = source["version"];
-	        this.releaseUrl = source["releaseUrl"];
+	        this.network = source["network"];
 	    }
 	}
 
@@ -378,6 +303,81 @@ export namespace config {
 	        this.vegaHome = source["vegaHome"];
 	        this.defaultNetwork = source["defaultNetwork"];
 	        this.telemetry = this.convertValues(source["telemetry"], TelemetryConfig);
+	    }
+	
+		convertValues(a: any, classs: any, asMap: boolean = false): any {
+		    if (!a) {
+		        return a;
+		    }
+		    if (a.slice) {
+		        return (a as any[]).map(elem => this.convertValues(elem, classs));
+		    } else if ("object" === typeof a) {
+		        if (asMap) {
+		            for (const key of Object.keys(a)) {
+		                a[key] = new classs(a[key]);
+		            }
+		            return a;
+		        }
+		        return new classs(a);
+		    }
+		    return a;
+		}
+	}
+
+}
+
+export namespace jsonrpc {
+	
+	export class Request {
+	    jsonrpc: string;
+	    method: string;
+	    params?: any;
+	    id?: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new Request(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.jsonrpc = source["jsonrpc"];
+	        this.method = source["method"];
+	        this.params = source["params"];
+	        this.id = source["id"];
+	    }
+	}
+	export class ErrorDetails {
+	    code: number;
+	    message: string;
+	    data?: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new ErrorDetails(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.code = source["code"];
+	        this.message = source["message"];
+	        this.data = source["data"];
+	    }
+	}
+	export class Response {
+	    jsonrpc: string;
+	    result?: any;
+	    error?: ErrorDetails;
+	    id?: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new Response(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.jsonrpc = source["jsonrpc"];
+	        this.result = source["result"];
+	        this.error = this.convertValues(source["error"], ErrorDetails);
+	        this.id = source["id"];
 	    }
 	
 		convertValues(a: any, classs: any, asMap: boolean = false): any {
